Move all classes into one package tracker
