+++ src/main/java/tracker/Command.java
@@ -1,3 +1,5 @@
+package tracker;
+
 public abstract class Command {
     public abstract boolean execute(TaskList taskList, Ui ui, Storage storage) throws TrackerException;
 }
