+++ src/main/java/tracker/Event.java
@@ -1,3 +1,5 @@
+package tracker;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
