+++ src/main/java/tracker/ExitCommand.java
@@ -1,3 +1,5 @@
+package tracker;
+
 public class ExitCommand extends Command {
     @Override
     public boolean execute(TaskList taskList, Ui ui, Storage storage) {
