+++ src/main/java/tracker/Task.java
@@ -1,3 +1,5 @@
+package tracker;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
