+++ src/main/java/tracker/TaskList.java
@@ -1,3 +1,5 @@
+package tracker;
+
 import java.util.ArrayList;
 
 public class TaskList {
@@ -17,7 +19,7 @@ public class TaskList {
 
     public Task removeTask(int index) throws TrackerException {
         if (index < 0 || index >= tasks.size()) {
-            throw new TrackerException("Task index out of range. Please check the list of tasks again.");
+            throw new TrackerException("tracker.Task index out of range. Please check the list of tasks again.");
         }
 
         return tasks.remove(index);
@@ -25,7 +27,7 @@ public class TaskList {
 
     public Task getTask(int index) throws TrackerException {
         if (index < 0 || index >= tasks.size()) {
-            throw new TrackerException("Task index out of range. Please check the list of tasks again.");
+            throw new TrackerException("tracker.Task index out of range. Please check the list of tasks again.");
         }
 
         return tasks.get(index);
