+++ src/main/java/tracker/Todo.java
@@ -1,3 +1,5 @@
+package tracker;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
