+++ src/main/java/tracker/TrackerException.java
@@ -1,3 +1,5 @@
+package tracker;
+
 public class TrackerException extends Exception {
     public TrackerException(String message) {
         super(message);
