+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-    Hello! I'm Tracker
+    Hello! I'm tracker.Tracker
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
